Extract shared SSM helpers into ingest_wikimedia/ssm.py
Both wikimedia_upload_status.py and wikimedia_launch.py contained
identical INSTANCE_ID/REGION constants and near-identical ssm_run()
implementations. Extract them into a shared module.

The canonical implementation is the improved poll loop from
wikimedia_launch.py (sleep before subsequent attempts, not after
InvocationDoesNotExist) with SSM_MAX_POLLS=60 (5 min) to cover
slow git clone + uv sync during EC2 code updates.
