+++ scripts/wikimedia_upload_status.py
@@ -5,49 +5,18 @@ Runs as a GitHub Action on a schedule and on workflow_dispatch (triggered by
 the /wikimedia-status Slack slash command via Lambda).
 """
 
-import json
 import logging
 import os
 import shlex
-import time
 from concurrent.futures import ThreadPoolExecutor, as_completed
 
 import boto3
 import requests
 
-INSTANCE_ID = "i-033eff6c8c168f999"
+from ingest_wikimedia.ssm import REGION, ssm_run
+
 SLACK_CHANNEL = "C02HEU2L3"
 SLACK_API_URL = "https://slack.com/api/chat.postMessage"
-REGION = "us-east-1"
-SSM_POLL_INTERVAL = 5
-SSM_MAX_POLLS = 24  # 2 minutes
-
-
-def ssm_run(client, cmd: str) -> str:
-    resp = client.send_command(
-        InstanceIds=[INSTANCE_ID],
-        DocumentName="AWS-RunShellScript",
-        Parameters={"commands": [f"sudo -u ec2-user bash -c {json.dumps(cmd)}"]},
-    )
-    cmd_id = resp["Command"]["CommandId"]
-    for _ in range(SSM_MAX_POLLS):
-        try:
-            inv = client.get_command_invocation(
-                CommandId=cmd_id, InstanceId=INSTANCE_ID
-            )
-        except client.exceptions.InvocationDoesNotExist:
-            time.sleep(SSM_POLL_INTERVAL)
-            continue
-        status = inv["Status"]
-        if status == "Success":
-            return inv.get("StandardOutputContent", "").strip()
-        if status in ("Failed", "TimedOut", "Cancelled"):
-            stderr = inv.get("StandardErrorContent", "").strip()
-            raise RuntimeError(
-                f"SSM command {cmd_id} ended with {status}: {stderr or 'no stderr'}"
-            )
-        time.sleep(SSM_POLL_INTERVAL)
-    raise TimeoutError(f"SSM command {cmd_id} did not complete within polling window")
 
 
 def get_phase_and_progress(client, partner: str) -> str:
